requirements.txt
pyyaml
seaborn
nvidia-ml-py3